requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
google-genai==1.2.0
pinecone==7.0.0
pypdf==5.0.0
python-docx==1.1.2
